front - review fixes
